+++ earth2studio/data/__init__.py
@@ -21,6 +21,7 @@ from .cams import CAMS_FX
 from .cbottle import CBottle3D
 from .cds import CDS
 from .cfs import CFS_FX, CFS_FX_Flux
+from .cfs_reforecast import CFS_Reforecast_FX, CFS_Reforecast_FX_Flux
 from .cmip6 import CMIP6, CMIP6MultiRealm
 from .const import Constant, Constant_FX
 from .ecmwf import AIFS_ENS_FX, AIFS_FX, IFS, IFS_ENS, IFS_ENS_FX, IFS_FX